requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
googlesearch-python==1.2.3
arxiv==1.4.8
pyalex==0.13
biopython==1.81
SPARQLWrapper==2.0.0
wikipediaapi==0.6.0
lxml==4.9.3